Add a test file
As noted in the README, this file cannot be used inside this repository.
However, the README points to a repository in which it can be used.

Currently, test #4 fails for an unknown reason. Following the same
commands from test #1 to #4 sequentially actually works right. So there
must be something different in the testing environment that makes it
fail.

Signed-off-by: Gabriel <lelutin@gmail.com>